Relocate style-components definitions
